+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.4.0 (unreleased)
 **Updated dependencies**
 - `kubernetes`: bumped to version 17.17
+- added scheduler options to adjust pod namespace and to allow scheduling of pods without a corresponding `SchedulerName` spec
 
 # v0.3.0 (2021-01-21)
 **Updated dependencies**