+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.4.0 (unreleased)
+# v0.4.0 (2021-10-27)
 **Updated dependencies**
 - `kubernetes`: bumped to version 19.15
 - added scheduler options to adjust pod namespace, allow scheduling of pods without a corresponding `SchedulerName` spec, and provide default CPU and memory resource requests for existing pods